Handling search error added

+++ update.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+# Exit immediately if a command fails
+set -e
+
 # Define variables for script and output file names
 SCRIPT1="sslbl_netlas_search.py"
 SCRIPT2="update_malicious_hosts_list.py"
@@ -12,12 +15,27 @@ else
     python3 "$SCRIPT1" -o "$OUTPUT_FILE" -a "$NETLAS_API_KEY" -q
 fi
 
-# Check if the output file exists after the first script completes
+# Check if the first script exited successfully
+if [[ $? -ne 0 ]]; then
+    echo "Error: $SCRIPT1 failed. Exiting."
+    exit 1
+fi
+
+# Check if the output file exists after the first script completes successfully
 if [[ -f "$OUTPUT_FILE" ]]; then
-    # Updating database script
+    
+    # Run the second script to update the database
     python3 "$SCRIPT2" "$OUTPUT_FILE"
-    # Delete the output file after the second script completes
+
+    # Check if the second script exited successfully
+    if [[ $? -ne 0 ]]; then
+        echo "Error: $SCRIPT2 failed. Exiting."
+        exit 2
+    fi
+
+    # Delete the output file after the second script completes successfully
     rm "$OUTPUT_FILE"
 else
     echo "Output file $OUTPUT_FILE not found. Second script will not run."
+    exit 2
 fi
